@@ -1,28 +1,67 @@
 {
   "name": "google-map-react",
   "version": "1.1.7",
-  "description": "isomorphic google map react component, allows render react components on the google map",
-  "main": "lib/index",
-  "scripts": {
-    "build:lib": "babel ./src -d lib",
-    "build:umd": "webpack src/index_umd.js dist/GoogleMapReact.js --config webpack.config.dev.js",
-    "build:umd:min": "webpack src/index_umd.js dist/GoogleMapReact.min.js --config webpack.config.prod.js",
-    "build": "yarn run build:lib && yarn run build:umd && yarn run build:umd:min",
-    "clean": "rimraf lib dist",
-    "prepublish": "yarn run clean && yarn run build",
-    "format": "prettier --trailing-comma es5 --single-quote --write 'src/**/*.js' 'src/**/*.js'",
-    "format:dev": "prettier --trailing-comma es5 --single-quote --write 'develop/**/*.js' 'develop/**/*.js'",
-    "lint": "eslint src",
-    "lint:dev": "eslint develop",
-    "test": "NODE_ENV=eee mocha --compilers js:babel-register --recursive --require babel-polyfill",
-    "test:watch": "NODE_ENV=eee mocha --compilers js:babel-register --recursive  --require babel-polyfill --watch",
-    "start": "kotatsu serve --port 4000 --config ./develop/config/index.babel.js --presets es2015,stage-0,react,react-hmre ./develop/Main.js",
-    "precommit": "lint-staged"
-  },
+  "description": "Isomorphic component that allows rendering react components on a google map",
+  "author": "istarkov https://github.com/istarkov",
+  "license": "MIT",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/google-map-react/google-map-react.git"
   },
+  "main": "dist/index.js",
+  "module": "dist/index.modern.js",
+  "source": "src/index.js",
+  "engines": {
+    "node": ">=10"
+  },
+  "scripts": {
+    "build": "microbundle-crl --no-compress --format modern,cjs",
+    "start": "microbundle-crl watch --no-compress --format modern,cjs",
+    "prepare": "run-s build",
+    "test": "run-s test:unit test:lint test:build",
+    "test:build": "run-s build",
+    "test:lint": "eslint .",
+    "test:unit": "cross-env CI=1 react-scripts test --env=jsdom",
+    "test:watch": "react-scripts test --env=jsdom",
+    "predeploy": "cd example && yarn install && yarn run build",
+    "deploy": "gh-pages -d example/build"
+  },
+  "dependencies": {
+    "@mapbox/point-geometry": "^0.1.0",
+    "eventemitter3": "^4.0.4",
+    "prop-types": "^15.7.2",
+    "scriptjs": "^2.5.9"
+  },
+  "peerDependencies": {
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
+  },
+  "devDependencies": {
+    "babel-eslint": "^10.0.3",
+    "cross-env": "^7.0.2",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.7.0",
+    "eslint-config-standard": "^14.1.0",
+    "eslint-config-standard-react": "^9.2.0",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-node": "^11.0.0",
+    "eslint-plugin-prettier": "^3.1.1",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.17.0",
+    "eslint-plugin-standard": "^4.0.1",
+    "expect": "1.20.2",
+    "jsdom": "^6.5.1",
+    "gh-pages": "^2.2.0",
+    "microbundle-crl": "^0.13.10",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.0.4",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-scripts": "^3.4.1"
+  },
+  "files": [
+    "dist"
+  ],
   "keywords": [
     "react",
     "reactjs",
