@@ -1,1 +1,5 @@
-export default from './google_map';
+import GoogleMap from './google_map';
+
+export * from './lib';
+
+export default GoogleMap;
