@@ -1,4 +1,4 @@
-const isEmpty = val => {
+const isEmpty = (val) => {
   // check for empty object {}, array []
   if (val !== null && typeof val === 'object') {
     if (Object.keys(val).length === 0) {
