@@ -6,6 +6,8 @@ const objectToString = Object.prototype.toString;
 
 export default function isNumber(value) {
   const numberTag = '[object Number]';
-  return typeof value === 'number' ||
-    (isObjectLike(value) && objectToString.call(value) === numberTag);
+  return (
+    typeof value === 'number' ||
+    (isObjectLike(value) && objectToString.call(value) === numberTag)
+  );
 }
