@@ -1,5 +1,5 @@
 // source taken from https://github.com/rackt/redux/blob/master/src/utils/isPlainObject.js
-const fnToString = fn => Function.prototype.toString.call(fn);
+const fnToString = (fn) => Function.prototype.toString.call(fn);
 
 /**
  * @param {any} obj The object to inspect.
@@ -10,9 +10,10 @@ export default function isPlainObject(obj) {
     return false;
   }
 
-  const proto = typeof obj.constructor === 'function'
-    ? Object.getPrototypeOf(obj)
-    : Object.prototype;
+  const proto =
+    typeof obj.constructor === 'function'
+      ? Object.getPrototypeOf(obj)
+      : Object.prototype;
 
   if (proto === null) {
     return true;
@@ -20,7 +21,9 @@ export default function isPlainObject(obj) {
 
   const constructor = proto.constructor;
 
-  return typeof constructor === 'function' &&
+  return (
+    typeof constructor === 'function' &&
     constructor instanceof constructor &&
-    fnToString(constructor) === fnToString(Object);
+    fnToString(constructor) === fnToString(Object)
+  );
 }
